init_ceremony/requirements.txt
kubernetes
validators
dnspython
